Resolve JAXB naming conflicts in tests
- Move test GuidelineCheck implementation to separate testutil package
- Create IOTestGuidelineCheck and IOTestScanReport classes to avoid naming conflicts
- Ensure all JAXB-annotated test classes have unique XML type names
- Refactor GuidelineIOTest to use the new test utility classes
